made login a separate window
